abstract background tasks, use one class and one function to retrieve API data from search words
